delete_db_rows2

+++ page_history.php
@@ -4,27 +4,31 @@
 </div>
 
 <?php
+$delete = $_POST['delete'];
 
+if($delete == 1){
 
-//DB DELETE
-try {
-    //DB接続
-    include("db_connect.php");
+    //DB DELETE
+    try {
+        //DB接続
+        include("db_connect.php");
 
-    // DELETE文を変数に格納
-    $sql = "DELETE FROM cart";
-    //削除する値は空のまま、SQL実行の準備をする
-    $stmt = $pdo->prepare($sql);
-    //executeにセットしてSQLを実行
-    $stmt->execute();
+        // DELETE文を変数に格納
+        $sql = "DELETE FROM cart";
+        //削除する値は空のまま、SQL実行の準備をする
+        $stmt = $pdo->prepare($sql);
+        //executeにセットしてSQLを実行
+        $stmt->execute();
 
-    echo "削除しました";
-} catch (PDOException $e) {
-    exit('データベースに接続できませんでした。' . $e->getMessage());
+        echo "削除しました";
+    } catch (PDOException $e) {
+        exit('データベースに接続できませんでした。' . $e->getMessage());
+    }
+
+    //変数リセット
+    $delete = "";
 }
-
 ?>
-
 <?php
 
 //DB output cart
